Update bleeding.lua
Moved random out of each repetition to have easier control over it. If we roll between 1 and 100, the check will never pass on 100 as you will get "lucky" with fresh roll on lower number. If we set the roll and keep it, we should actually have to wait till 100 if we roll it.

This is step 1 for tweaking numbers.