Bleeding option (#35)
* Add files via upload

* Add files via upload

* Delete twitch_fragments/outcomes/oily.lua

* Bleeding outcome

* Delete files/pixel_scenes/oily.png

* Bleeding outcome

* Update bleeding.lua

* Update bleeding.lua

Moved random out of each repetition to have easier control over it. If we roll between 1 and 100, the check will never pass on 100 as you will get "lucky" with fresh roll on lower number. If we set the roll and keep it, we should actually have to wait till 100 if we roll it.

This is step 1 for tweaking numbers.

* Update bleeding.lua

Reduce the average stuns down to 4, but there can be stuns in two bleeding intervals in a row. Very unpredictable for the player.

* Update bleeding.lua

Same category as take dmg and electrocution/

* Update manifest.json

I download file from github (as it sometimes changes end of line character, so local files can be slightly different), generate hash with https://www.strerr.com/en/sha256_file.html or similar method and put the sha in manifest.json so it can be downloaded on TI update

---------

Co-authored-by: Miczu <48890028+Miczu@users.noreply.github.com>