The Immortal Snail
New vote option: Immortal Snail
Very rare vote, adds the immortal snail to the game permanently when voted in.
If the vote occurs multiple times in the same runs, multiple snails can be present.
The snail will steal perks and items if it manages to touch them

README: I was able to test this in a separate mod and have confirmed it worked there but still can't get TI to load properly to confirm it works with the actual mod.
It should work without issues, but please double check this event functions properly before merging.

Unnecessary snail lore:
For those unfamiliar with the immortal snail, the immortal snail originates from a thought experiment asking what you would do if you were to receive a large amount of money, but in exchange an immortal snail aware of your location pursued you until the ends of the earth and if it touches you at any time, you would immediately die.

Some variations of the thought experiment involve the user being immortal or other tweaks, but the important part is this concept was added to a mod called FairMod.

In Fairmod the immortal snail manifests as something which will pursue the player indefinitely, will always be aware of their location and will immediately kill the player upon making contact with them.

The snail will keep up with the player by teleporting to a location nearby them if they stray too far away and in true fairmod spirit, the snail will steal any peaks or items it makes contact with to use them against the player. (This wasn't originally an intended feature but ended up making its way in and it was way too funny to leave out lmao)

If you still aren't sure what the immortal snail is, just watch Dunks fairmod videos and you'll get the idea.
Just like fairmod, this PR description is unnecessarily convoluted.